Create Day 2 - Tip Calculator.py
A tip calculator which asks the user for the total bill, the tip percentage they would like to give, and how many people are going to split the bill. The amount to be payed is calculated by multiplying the total bill by the tip as a percentage plus 1, and then dividing that amount by the number of people. The amount to be payed is then formatted to display two digits after the decimal point.