Create escape_the_maze.py
This code is for the escape the Maze game on Reeborg's World. The object of the game is to write code which will enable the Robot to navigate through the maze and find the exit. Every time the game starts, the position of the robot as well as the direction the robot is facing changes randomly so the code has to work no matter where the robot is and which direction it is facing.
The link for this can be found here: https://reeborg.ca/reeborg.htmllang=en&mode=python&menu=worlds%2Fmenus%2Freeborg_intro_en.json&name=Maze&url=worlds%2Ftutorial_en%2Fmaze1.json