Create hurdle 4.py
This code is for the Reeborg's World Hurdle 4. The object of the game is to write code which enables the Robot to clear the hurdles and travel to the finish line. Every time the game starts, the position and height of the hurdles changes randomly so the code has to work no matter what the hurdles look like. The link to the game can be found here: https://reeborg.ca/reeborg.html?lang=en&mode=python&menu=worlds%2Fmenus%2Freeborg_intro_en.json&name=Hurdle%204&url=worlds%2Ftutorial_en%2Fhurdle4.json