Create Day 7 - Hangman
A program for the Hangman game where the goal is for the user to correctly guess all the letters in a word. The user starts out with six lives and must correctly guess a letter in order to not lose a life. In this program, a word is picked randomly from a list of words in the hangman_words.py file.  A list containing blank spaces ("_") is created which will have the correctly guessed letters added to it and will be he used to display them to the user as the game progresses. If the user guesses an incorrect letter, the user loses a life and the next stage of the hangman game will be display. The ascii art for this is in the hangman_art.py file. The program also lets the user know if they've already guessed a correct letter and doesn't count this as a guess. If all of the blank spaces have been filled and the user still has at least 1 life, then they win.