Create hangman.py
A program for the Hangman game where the object of the game is for the user to correctly guess all the letters in a word. At the start of the game, the user has 6 lives and a word from the list in the hangman_words.py file is randomly selected. A list which contain blank spaces ("_") is created which will have each correctly guessed letter added to it at the correct index position, and will be used to display the correct letters to the users. If the user incorrectly guesses a letter, they will lose a life and the appropriate "stage" image from the ascii art in the hangman_art.py file will be displayed. The program will also tell the user if they're already guessed a correct letter and will not punish them for this. If the user guessed a incorrect letter that they have already guessed, it will count this as a guess and they will lose a life. If the list containing the correctly guessed letters no longer contains any blank spaces ("_") and the user still has at least one life, they win the game.