Create secret_auction.py
A secret auction program coded on https://replit.com/~ in order to use the built in clear() function. The program asks the user for their name, how much they want to bid, and if there is another bidder. If the user answers 'yes', then the screen is cleared before the next bidder  is prompted to enter their information. If the user answer's 'no', then a message is printed revealing the name of the person who bid the highest and the bid amount.