Create blackjack.py
A Blackjack game coded on replit.com in order to use the built in clear() function.  In my version of the game, each card has the numeric value of the number printed on it ( 5 has the value of 5, 7 has the value of 7, etc) but the King, Queen, and Joker cards have the value of 10. That's why there are four number 10s in the cards list. The number 11 in the cards list represents the Ace card. It initially has the value of 11 but will change to 1 if either the dealer or the player have the Ace in their hand and their score is over 21 (see lines 20 - 22). My version of the game also does not remove a card from the cards list once it has been drawn, therefore the game uses an infinite deck. The object of the game is for each player to get a score as close to 21 as possible, without going over 21.