Create guessing_game.py
A Guessing Game program which generates a random number between 1 and 100 and the goal of the game is for the user to correctly guess that number. There are two game modes for the user to choose from - easy and hard. If easy is chosen, the user has 10 attempts to guess the number and if hard is chosen, the user has 5 attempts. The game will also tell the player if their guesses are higher or lower than the number. If the user guesses correctly they win but if not and they run out of lives they lose.