Create coffee_machine.py
A coffee machine program which gives the user three choices - espresso, latte, and cappuccino and allows the user to pay in pennies, nickels, dimes, and quarters. It determines if the users has provided enough money or not, if so it calculates the change for the user. It also keeps track of how much water, milk, and coffee has been used and if there is not enough of one or more of these resources, it cannot fulfil any more orders. It also generates reports displaying the quantities of resources and the amount of money earned.