Create turtle_race.py
A program that creates a racing game using the Turtle library. There are 6 differently coloured turtles in the race and the user is asked to bet which one will win. Each turtle then races across the screen to the finish line and console displays whether the user bet correctly or not.