Create main.py
The second part of the program for the classic Snake game. Code has been added to detect collision with the food, the walls, and the Snake's tail, as well as a scorekeeping mechanic and a scoreboard disapply. The code for the program has also been split up into different classes.