Create main.py
The code to create an invitation letter for each name in the invited_names.txt. The program opens the invited_names.txt file and stores it as a list in the names variable. Then the starting_letter.txt file (which is the template for the invitation letter) is opened and stored as a string in the letter_contents variable. A for loop is used to iterate through each name in the names list to strip the new lines between each name, and replace the [name] placeholder with the name of the intended recipient. The completed invitation letters are the created and written to.