Update main.py
A Password Manager App which allows the user to save login details for different websites.  The user can either enter a password themselves or click the Generate Password button and a random password is created for them. When the user clicks the Add button, the login details are saved to a text file.