Create main.py
A program which alerts the user of fluctuations in price for Tesla stock. It compares yesterday's closing price to the day before yesterday and if there has been a change of 5% or more, the user will be sent three SMSes. Each SMS displays the name of the stock and the difference in price and contains a news article pertaining to Tesla Inc. The program makes use of the AlphaVantage API to get stock information and the NewsAPI to get news articles.